+++ examples/volumetric/slicer1.py
@@ -3,22 +3,18 @@ Click button to change colormap"""
 from vedo import dataurl, Volume, Text2D
 from vedo.applications import Slicer3DPlotter
 
-filename = dataurl + "embryo.slc"
-# filename = dataurl+'embryo.tif'
-# filename = dataurl+'vase.vti'
-
-vol = Volume(filename)#.print()
+vol = Volume(dataurl + "embryo.slc")
 
 plt = Slicer3DPlotter(
     vol,
-    bg="white",
-    bg2="lightblue",
     cmaps=("gist_ncar_r", "jet", "Spectral_r", "hot_r", "bone_r"),
     use_slider3d=False,
+    bg="white",
+    bg2="lightblue",
 )
 
-# Can now add any other object to the Plotter scene:
-# plt += Text2D('some message')
-# plt.show().interactive()
+# Can now add any other vedo object to the Plotter scene:
+plt += Text2D(__doc__)
 
+plt.show()
 plt.close()